+++ rich/console.py
@@ -5,6 +5,7 @@ import shutil
 import sys
 import threading
 from abc import ABC, abstractmethod
+from collections import abc
 from dataclasses import dataclass, field
 from datetime import datetime
 from functools import wraps
@@ -997,7 +998,7 @@ class Console:
 
     def status(
         self,
-        status: TextType,
+        status: RenderableType,
         *,
         spinner: str = "dots",
         spinner_style: str = "status.spinner",
@@ -1007,7 +1008,7 @@ class Console:
         """Display a status and spinner.
 
         Args:
-            status (TextType): Text for a given status.
+            status (RenderableType): A status renderable (str or Text typically).
             console (Console, optional): Console instance to use, or None for global console. Defaults to None.
             spinner (str, optional): Name of spinner animation (see python -m rich.spinner). Defaults to "dots".
             spinner_style (StyleType, optional): Style of spinner. Defaults to "status.spinner".